Add solution for day 4
Silly optimization of using a byte as the set of fields seen. Each field
corresponds to a single bit mask:

  byr: 64 -> 01000000
  iyr: 32 -> 00100000
  eyr: 16 -> 00010000
  hgt:  8 -> 00001000
  hcl:  4 -> 00000100
  ecl:  2 -> 00000010
  pid:  1 -> 00000001
  cid:  0 -> 00000000

After we parse each field of an entry into its mask, we bitwise-or
them together. If we get:

      127 -> 01111111

then all fields are present. For Part2, we just do some extra validation
of the field's value and return 0 for the mask (treating it as
not-present) if the value is malformed.