Add solution for day 7
The input is an adjacency list, which can be parsed into a graph
represented by:

  HashMap Text [(Int, Text)]

For part 1, we invert the graph and then count the unique "descendents"
reachable from the root ("shiny gold").

For part 2, we walk the input graph from the root and sum the result of
multiplying each child's count by the recursive step.

Both parts can be decomposed into this procedure using a different
monoid, which is kind of neat:

  foldDescendents
    :: Monoid m
    => (Text -> m)
    -> (Int -> m -> m)
    -> Text
    -> Graph
    -> m
  foldDescendents wrap step root graph = without root
   where
    with label = wrap label <> without label
    without label = mconcat $ do
      children <- maybeToList $ HashMap.lookup label graph
      (n, child) <- children
      pure $ step n $ with child

The mutual recursion between `with` and `without` lets us ignore the
root node without some special case.

Part 1 uses set union:

  uniqueDescendents :: Text -> Graph -> HashSet Text
  uniqueDescendents = foldDescendents wrap step
   where
    wrap = HashSet.singleton
    step _ acc = acc

Part 2 uses sum:

  countDescendents :: Text -> Graph -> Int
  countDescendents root = getSum . foldDescendents wrap step root
   where
    wrap = const 1
    step !n !acc = Sum n * acc